+++ src/main/java/dev/willyelton/crystal_tools/client/renderer/CrystalToolsRenderTypes.java
@@ -4,6 +4,7 @@ import com.mojang.blaze3d.pipeline.BlendFunction;
 import com.mojang.blaze3d.pipeline.RenderPipeline;
 import com.mojang.blaze3d.vertex.DefaultVertexFormat;
 import com.mojang.blaze3d.vertex.VertexFormat;
+import com.mojang.blaze3d.vertex.VertexFormatElement;
 import dev.willyelton.crystal_tools.CrystalTools;
 import net.minecraft.client.renderer.RenderPipelines;
 import net.minecraft.client.renderer.RenderStateShard;
@@ -11,9 +12,10 @@ import net.minecraft.client.renderer.RenderType;
 import net.minecraft.resources.ResourceLocation;
 
 import static dev.willyelton.crystal_tools.CrystalTools.rl;
-import static net.minecraft.client.renderer.RenderPipelines.ENTITY_SNIPPET;
+import static net.minecraft.client.renderer.RenderPipelines.MATRICES_FOG_SNIPPET;
 import static net.minecraft.client.renderer.RenderPipelines.MATRICES_PROJECTION_SNIPPET;
-import static net.minecraft.client.renderer.RenderStateShard.*;
+import static net.minecraft.client.renderer.RenderStateShard.NO_LIGHTMAP;
+import static net.minecraft.client.renderer.RenderStateShard.OVERLAY;
 import static net.minecraft.client.renderer.RenderType.create;
 
 public class CrystalToolsRenderTypes {
@@ -31,15 +33,23 @@ public class CrystalToolsRenderTypes {
             .withLocation(rl("pipeline/block_overlay"))
             .build();
 
-    public static RenderPipeline QUARRY_CUBE_PIPELINE = RenderPipeline.builder(ENTITY_SNIPPET)
+    public static final VertexFormat POSITION_TEX_LIGHTMAP_COLOR = VertexFormat.builder()
+            .add("Position", VertexFormatElement.POSITION)
+            .add("UV0", VertexFormatElement.UV0)
+            .add("UV2", VertexFormatElement.UV2)
+            .add("Color", VertexFormatElement.COLOR)
+            .build();
+
+    public static RenderPipeline QUARRY_CUBE_PIPELINE = RenderPipeline.builder(MATRICES_FOG_SNIPPET)
             .withLocation(rl("pipeline/quarry_cube"))
-            .withVertexShader("core/entity")
-            .withFragmentShader("core/entity")
-            .withShaderDefine("ALPHA_CUTOUT", 0.1F)
-            .withSampler("Sampler1")
-            .withBlend(BlendFunction.TRANSLUCENT)
+//            .withShaderDefine("ALPHA_CUTOUT", 0.1F)
+//            .withShaderDefine("PER_FACE_LIGHTING")
+            .withSampler("Sampler0")
+            .withSampler("Sampler2")
+            .withVertexShader(rl("position_color_tex_lightmap"))
+            .withFragmentShader(rl("position_color_tex_lightmap"))
+            .withVertexFormat(DefaultVertexFormat.POSITION_COLOR_TEX_LIGHTMAP, VertexFormat.Mode.QUADS)
             .withCull(false)
-            .withVertexFormat(DefaultVertexFormat.NEW_ENTITY, VertexFormat.Mode.QUADS)
             .build();
 
     // Copied from the DEBUG_QUADS Render Type
@@ -65,9 +75,9 @@ public class CrystalToolsRenderTypes {
             1536,
             true,
             false,
-            QUARRY_CUBE_PIPELINE,
+            RenderPipelines.ENTITY_CUTOUT_NO_CULL,
             RenderType.CompositeState.builder()
-                    .setTextureState(new RenderStateShard.TextureStateShard(ResourceLocation.fromNamespaceAndPath(CrystalTools.MODID, "textures/entity/crystal_quarry_cube.png"), false))
+                    .setTextureState(new RenderStateShard.TextureStateShard(rl("textures/entity/crystal_quarry_cube.png"), false))
                     .setLightmapState(NO_LIGHTMAP)
                     .setOverlayState(OVERLAY)
                     .createCompositeState(true));