+++ src/main/java/dev/willyelton/crystal_tools/client/renderer/blockentity/CrystalQuarryBlockEntityRenderer.java
@@ -1,9 +1,12 @@
 package dev.willyelton.crystal_tools.client.renderer.blockentity;
 
 import com.mojang.blaze3d.vertex.PoseStack;
+import dev.willyelton.crystal_tools.client.renderer.CrystalToolsRenderTypes;
 import dev.willyelton.crystal_tools.client.renderer.QuarryCubeModel;
+import dev.willyelton.crystal_tools.client.renderer.QuarryLaserRenderer;
 import dev.willyelton.crystal_tools.common.levelable.block.entity.CrystalQuarryBlockEntity;
 import dev.willyelton.crystal_tools.utils.Colors;
+import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.SubmitNodeCollector;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderer;
 import net.minecraft.client.renderer.blockentity.BlockEntityRendererProvider;
@@ -16,9 +19,10 @@ import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.ArrayList;
 import java.util.List;
 
-import static dev.willyelton.crystal_tools.client.renderer.CrystalToolsRenderTypes.QUARRY_CUBE;
+import static dev.willyelton.crystal_tools.CrystalTools.rl;
 
 public class CrystalQuarryBlockEntityRenderer implements BlockEntityRenderer<CrystalQuarryBlockEntity, CrystalQuarryBlockEntityRenderer.QuarryRenderState> {
     private static final int FRAME_COLOR = Colors.fromRGB(48, 231, 237, 255);
@@ -30,47 +34,6 @@ public class CrystalQuarryBlockEntityRenderer implements BlockEntityRenderer<Cry
         this.model = new QuarryCubeModel(context.bakeLayer(QuarryCubeModel.LOCATION));
     }
 
-//    @Override
-//    public void render(CrystalQuarryBlockEntity blockEntity, float partialTick, PoseStack poseStack, MultiBufferSource bufferSource, int packedLight, int packedOverlay, Vec3 cameraPos) {
-//        BlockPos miningAt = blockEntity.getMiningAt();
-//
-//        if (miningAt != null) {
-//            BlockPos quarryPos = blockEntity.getBlockPos();
-//            cubeRenderer.render((int) blockEntity.getLevel().getGameTime(), partialTick, poseStack, bufferSource, packedLight,
-//                    blockEntity.getCenterX() - quarryPos.getX(), blockEntity.getCenterY() - quarryPos.getY(), blockEntity.getCenterZ() - quarryPos.getZ());
-//
-//
-//            Camera camera = Minecraft.getInstance().gameRenderer.getMainCamera();
-//            Vec3 view = camera.getPosition();
-//            int color = Colors.fromRGB(48, 231, 237, 255);
-//            int miningLaserColor = Colors.fromRGB(0, 144, 180, 255);
-//
-//            poseStack.pushPose();
-//            // The renderer expects it in camera space, the block entity renderer is initially in block space
-//            poseStack.translate((double) -blockEntity.getBlockPos().getX() + view.x, (double) -blockEntity.getBlockPos().getY() + view.y, (double) -blockEntity.getBlockPos().getZ() + view.z);
-//
-//            float centerX = blockEntity.getCenterX();
-//            float centerY = blockEntity.getCenterY();
-//            float centerZ = blockEntity.getCenterZ();
-//            QuarryLaserRenderer.renderLaser(bufferSource, poseStack, camera, partialTick, blockEntity.getLevel(), centerX, centerY, centerZ, blockEntity.getMiningAt(), miningLaserColor);
-//
-//            List<BlockPos> corners = blockEntity.getStabilizerPositions();
-//
-//            if (!corners.isEmpty()) {
-//                QuarryLaserRenderer.renderLaser(bufferSource, poseStack, camera, partialTick, blockEntity.getLevel(), corners.get(0), corners.get(1), color);
-//                QuarryLaserRenderer.renderLaser(bufferSource, poseStack, camera, partialTick, blockEntity.getLevel(), corners.get(1), corners.get(2), color);
-//                QuarryLaserRenderer.renderLaser(bufferSource, poseStack, camera, partialTick, blockEntity.getLevel(), corners.get(2), corners.get(3), color);
-//                QuarryLaserRenderer.renderLaser(bufferSource, poseStack, camera, partialTick, blockEntity.getLevel(), corners.get(3), corners.get(0), color);
-//            }
-//
-//            for (BlockPos corner : corners) {
-//                QuarryLaserRenderer.renderLaser(bufferSource, poseStack, camera, partialTick, blockEntity.getLevel(), corner, centerX, centerY, centerZ, color);
-//            }
-//
-//            poseStack.popPose();
-//        }
-//    }
-
     @Override
     public QuarryRenderState createRenderState() {
         return new QuarryRenderState();
@@ -81,9 +44,29 @@ public class CrystalQuarryBlockEntityRenderer implements BlockEntityRenderer<Cry
         BlockEntityRenderer.super.extractRenderState(blockEntity, renderState, partialTick, cameraPos, overlay);
         BlockPos quarryPos = blockEntity.getBlockPos();
 
+        renderState.quarryPos = blockEntity.getBlockPos();
+        renderState.miningAt = blockEntity.getMiningAt();
         renderState.rotation = ((float) blockEntity.getLevel().getGameTime() + partialTick) * 3.0F;
+        renderState.cubePos = new Vec3(blockEntity.getCenterX() - quarryPos.getX(), blockEntity.getCenterY() - quarryPos.getY() - 1.5, blockEntity.getCenterZ() - quarryPos.getZ());
+        renderState.centerPos = new Vec3(blockEntity.getCenterX(), blockEntity.getCenterY(), blockEntity.getCenterZ());
+        renderState.ageInTicks = blockEntity.getLevel().getGameTime() + partialTick;
+        List<QuarryLaserRenderer.QuarryLaserRenderState> laserRenderStates = new ArrayList<>();
+        laserRenderStates.add(QuarryLaserRenderer.extractQuarryLaserRenderState(renderState.centerPos, blockEntity.getMiningAt().getCenter(), cameraPos, blockEntity.getLevel(), partialTick, MINING_COLOR));
 
-        renderState.cubePos = new Vec3(blockEntity.getCenterX() - quarryPos.getX(), blockEntity.getCenterY() - quarryPos.getY(), blockEntity.getCenterZ() - quarryPos.getZ());
+        List<BlockPos> corners = blockEntity.getStabilizerPositions();
+
+        if (corners.size() == 4) {
+            laserRenderStates.add(QuarryLaserRenderer.extractQuarryLaserRenderState(corners.get(0), corners.get(1), cameraPos, blockEntity.getLevel(), partialTick, FRAME_COLOR));
+            laserRenderStates.add(QuarryLaserRenderer.extractQuarryLaserRenderState(corners.get(1), corners.get(2), cameraPos, blockEntity.getLevel(), partialTick, FRAME_COLOR));
+            laserRenderStates.add(QuarryLaserRenderer.extractQuarryLaserRenderState(corners.get(2), corners.get(3), cameraPos, blockEntity.getLevel(), partialTick, FRAME_COLOR));
+            laserRenderStates.add(QuarryLaserRenderer.extractQuarryLaserRenderState(corners.get(3), corners.get(0), cameraPos, blockEntity.getLevel(), partialTick, FRAME_COLOR));
+        }
+
+        for (BlockPos corner : corners) {
+            laserRenderStates.add(QuarryLaserRenderer.extractQuarryLaserRenderState(corner.getCenter(), renderState.centerPos, cameraPos, blockEntity.getLevel(), partialTick, FRAME_COLOR));
+        }
+
+        renderState.quarryLaserRenderStates = laserRenderStates;
     }
 
     @Override
@@ -94,7 +77,8 @@ public class CrystalQuarryBlockEntityRenderer implements BlockEntityRenderer<Cry
             BlockPos quarryPos = renderState.quarryPos;
             poseStack.pushPose();
             poseStack.translate(renderState.cubePos);
-            nodeCollector.submitModel(model, renderState, poseStack, QUARRY_CUBE, renderState.lightCoords, OverlayTexture.NO_OVERLAY, -1, null);
+            nodeCollector.submitModel(model, renderState, poseStack, RenderType.entityCutoutNoCull(rl("textures/entity/crystal_quarry_cube.png")), 0x00F000F0,
+                    OverlayTexture.NO_OVERLAY, 0, null);
             poseStack.popPose();
 
             Vec3 view = cameraRenderState.pos;
@@ -103,22 +87,11 @@ public class CrystalQuarryBlockEntityRenderer implements BlockEntityRenderer<Cry
             // The renderer expects it in camera space, the block entity renderer is initially in block space
             poseStack.translate((double) -quarryPos.getX() + view.x, (double) -quarryPos.getY() + view.y, (double) -quarryPos.getZ() + view.z);
 
-            float centerX = (float) renderState.centerPos.x();
-            float centerY = (float) renderState.centerPos.y();
-            float centerZ = (float) renderState.centerPos.z();
-//            QuarryLaserRenderer.renderLaser(bufferSource, poseStack, cameraRenderState, partialTick, blockEntity.getLevel(), centerX, centerY, centerZ, blockEntity.getMiningAt(), MINING_COLOR);
-
-            List<BlockPos> corners = renderState.stabilizerPositions;
-
-            if (!corners.isEmpty()) {
-//                QuarryLaserRenderer.renderLaser(bufferSource, poseStack, cameraRenderState, partialTick, blockEntity.getLevel(), corners.get(0), corners.get(1), FRAME_COLOR);
-//                QuarryLaserRenderer.renderLaser(bufferSource, poseStack, cameraRenderState, partialTick, blockEntity.getLevel(), corners.get(1), corners.get(2), FRAME_COLOR);
-//                QuarryLaserRenderer.renderLaser(bufferSource, poseStack, cameraRenderState, partialTick, blockEntity.getLevel(), corners.get(2), corners.get(3), FRAME_COLOR);
-//                QuarryLaserRenderer.renderLaser(bufferSource, poseStack, cameraRenderState, partialTick, blockEntity.getLevel(), corners.get(3), corners.get(0), FRAME_COLOR);
-            }
-
-            for (BlockPos corner : corners) {
-//                QuarryLaserRenderer.renderLaser(bufferSource, poseStack, cameraRenderState, partialTick, blockEntity.getLevel(), corner, centerX, centerY, centerZ, FRAME_COLOR);
+            for (QuarryLaserRenderer.QuarryLaserRenderState quarryLaserRenderState : renderState.quarryLaserRenderStates) {
+                nodeCollector.submitCustomGeometry(poseStack, CrystalToolsRenderTypes.QUARRY_LASER,
+                        (pose, consumer) -> {
+                            QuarryLaserRenderer.renderLaser(poseStack, consumer, quarryLaserRenderState);
+                        });
             }
 
             poseStack.popPose();
@@ -143,10 +116,11 @@ public class CrystalQuarryBlockEntityRenderer implements BlockEntityRenderer<Cry
     public static class QuarryRenderState extends BlockEntityRenderState {
         public BlockPos miningAt;
         public BlockPos quarryPos;
+        // Relative cube pos to block entity
         public Vec3 cubePos;
         public Vec3 centerPos;
         public float rotation;
-        public int ageInTicks;
-        public List<BlockPos> stabilizerPositions;
+        public float ageInTicks;
+        public List<QuarryLaserRenderer.QuarryLaserRenderState> quarryLaserRenderStates;
     }
 }